Rename package from eval to sdk.android
